--- polls/test.py
@@ -1,26 +1,0 @@
-from .models import Question
-import datetime
-
-from django.test import TestCase
-from django.utils import timezone
-
-
-class QustionModelTests(TestCase):
-
-    def test_was_published_recently_with_old_question(self):
-    """
-    was_published_recetly() returns False for question whose pub_date
-    is older than 1 day
-    """
-    time = timezone.now() - datetime.timedelta(days=1, seconds=1)
-    old_question = Question(pub_date=time)
-    self.assertIs(old_question.was_published_recently(), False)
-
-    def test_was_published_recently_with_future_question(self):
-    """
-    was_published_recently() returns True for questions whose pub_date
-    is within the last day.
-    """
-    time = timezone.now() + datetime.timedelta(days=30)
-    future_question = Question(pub_date=time)
-    self.assertIs(future_question.was_published_recently(), False)
